# api-docs.json
{
  "openapi": "3.1.0",
  "info": { "title": "DevSolve" },
  "servers": [
    {
      "url": "https://devsolve-api.quizzy.it.com",
      "description": "Generated server url"
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/user-profiles/me/avatar": {
      "put": {
        "tags": ["user-controller"],
        "operationId": "uploadAvatar",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserProfileResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["user-controller"],
        "operationId": "removeAvatar",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UserProfileResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organizations/register": {
      "post": {
        "tags": ["organization-controller"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrganizationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SocialLinkResponse": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "GITHUB",
              "LINKEDIN",
              "WEBSITE",
              "X",
              "FACEBOOK",
              "TELEGRAM",
              "OTHER"
            ]
          },
          "url": { "type": "string" }
        }
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "fullName": { "type": "string" },
          "biography": { "type": "string" },
          "phone": { "type": "string" },
          "avatarUrl": { "type": "string" },
          "dateOfBirth": { "type": "string", "format": "date" },
          "gender": { "type": "string", "enum": ["MALE", "FEMALE", "OTHER"] },
          "country": { "type": "string" },
          "socialLinks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SocialLinkResponse" }
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "SUSPENDED", "REMOVED"]
          },
          "reputation": { "type": "integer", "format": "int32" },
          "totalReports": { "type": "integer", "format": "int32" },
          "validReports": { "type": "integer", "format": "int32" },
          "criticalReports": { "type": "integer", "format": "int32" },
          "recognitionCount": { "type": "integer", "format": "int32" },
          "lastLoginAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "OrganizationRequest": {
        "type": "object",
        "properties": {
          "fullName": { "type": "string", "maxLength": 150, "minLength": 2 },
          "jobTitle": { "type": "string", "maxLength": 100, "minLength": 0 },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255,
            "minLength": 0
          },
          "password": { "type": "string", "maxLength": 100, "minLength": 8 },
          "confirmPassword": {
            "type": "string",
            "maxLength": 100,
            "minLength": 8
          },
          "companyName": { "type": "string", "maxLength": 255, "minLength": 0 },
          "companyWebsite": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "industry": {
            "type": "string",
            "enum": [
              "TECHNOLOGY",
              "FINANCE",
              "HEALTHCARE",
              "ECOMMERCE",
              "GOVERNMENT",
              "EDUCATION",
              "OTHER"
            ]
          },
          "companySize": {
            "type": "string",
            "minLength": 1,
            "pattern": "^(1-10|11-50|51-200|201-500|501-1000|1000\\+)$"
          },
          "country": { "type": "string", "maxLength": 100, "minLength": 0 },
          "joiningReason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          }
        },
        "required": [
          "companyName",
          "companySize",
          "companyWebsite",
          "confirmPassword",
          "country",
          "email",
          "fullName",
          "industry",
          "jobTitle",
          "joiningReason",
          "password"
        ]
      },
      "OrganizationResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "ownerId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "domain": { "type": "string" },
          "websiteUrl": { "type": "string" },
          "logoUrl": { "type": "string" },
          "description": { "type": "string" },
          "industry": {
            "type": "string",
            "enum": [
              "TECHNOLOGY",
              "FINANCE",
              "HEALTHCARE",
              "ECOMMERCE",
              "GOVERNMENT",
              "EDUCATION",
              "OTHER"
            ]
          },
          "companySize": { "type": "string" },
          "country": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["PENDING", "ACTIVE", "REJECTED"]
          },
          "submissionVersion": { "type": "integer", "format": "int32" },
          "rejectionReason": { "type": "string" },
          "reviewedAt": { "type": "string", "format": "date-time" },
          "verifiedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}